Increase receive buffer size to 40KB
Receiving data with small buffer sizes, like 1024 bytes as was currently the case, can cause the download of files to be come extremely slow on .NET Core on Linux.
We've seen timeouts when receiving an 8MB framebuffer.

This fix currently is to increase the buffer size to a more reasonable size.