Rewrite for when fields are missing from ADB output
When trying to parse the output from "adb devices -l" while my phone was in recovery mode, there was an "Invalid device list data" error because TWRP did not provide a model. I don't understand regex, so this is how I fixed it for my own personal use. This might need some more work, but it's good enough for my personal use.