Use Microsoft.Extensions.Logging for logging
This removes the need to use a custom logging framework in SharpAdbClient.

For compatibility with previous versions, all constructors default to a null value for the logger. This is converted to NullLogger<T>.

NullLogger<T> is only availably starting with Microsoft.Extensions.Logging version 2.0.0, which requires netstandard2.0.
As a result, support for .NET 4.5.1 has been dropped, and there's a version bump to 2.3.