Remove AdbClient.Instance
An AdbClient wraps around one or more a unique connections to an adb server. This connection is not intended to be shared; and storing it in a static can cause all kinds of mayhem (not in the least in unit tests).

This commit completely removes the AdbClient.Instance static.

Code which used to rely on AdbClient.Instance will now need to explictely pass the IAdbClient object. This makes it much more clear that you are operating on an AdbConnection.

For example:

```
device.ExecuteShellCommand(...)
```

now becomes
```
client.ExecuteShellCommand(device, ...)
```

making it much more explicit over which connection channel you are operating.